added logic CRUD for APS; checked relations before delete tree node

+++ src/Infrastructure/Server/SmartEcoAPI/Controllers/ASM/AreasController.cs
@@ -113,6 +113,8 @@ namespace SmartEcoAPI.Controllers.ASM
             {
                 return NotFound();
             }
+            if (await _context.AirPollutionSourceArea.AnyAsync(sw => sw.AreaId == area.Id))
+                return Forbid();
 
             _context.Area.Remove(area);
             await _context.SaveChangesAsync();
